add edit and destroy for few pages

+++ pages/team/team.php
@@ -8,7 +8,7 @@
 
 
 
-<?php $reponse = $pdo->query("SELECT professeur.nom, professeur.prenom, professeur.profession,professeur.description, professeur.mobile, professeur.email, image.url FROM professeur INNER JOIN image ON professeur.image_id = image.id" );
+<?php $reponse = $pdo->query("SELECT professeur.nom, professeur.prenom, professeur.profession,professeur.description, professeur.mobile, professeur.email, image.url, professeur.id FROM professeur INNER JOIN image ON professeur.image_id = image.id" );
   while ( $reponse_tableau = $reponse->fetch(PDO::FETCH_ASSOC)) {
    echo '<div>';
     echo $reponse_tableau["prenom"]."</br>";
@@ -18,15 +18,14 @@
     echo $reponse_tableau["mobile"]."</br>";
     echo $reponse_tableau["email"]."</br>";
     echo $reponse_tableau["url"]."</br>";
+    echo "<a href=\"edit.php?id=$reponse_tableau[id]\"><i class='fas fa-edit'></i></a>";
+    echo "<a href=\"delete.php?id=$reponse_tableau[id]\" onClick=\"return confirm('Are you sure you want to delete?')\"><i class='fas fa-trash'></i></a>";
     ?>
-    <a href="edit.php"><i class="fas fa-edit"></i></a>
-    <a href="delete.php"><i class="fas fa-trash"></i></a>
 
     <?php
        echo '</div>';
        echo '<hr>';
   }
 ?>
-
 
 <?php include "../../templates/footer.php"; ?>
